Use React as a template language in permission-boundaries
