+++ frontend/packages/aquascope-editor/package.json
@@ -11,7 +11,13 @@
     "@codemirror/view": "^6.4.0",
     "lodash": "^4.17.21"
   },
+  "peerDependencies": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
+  },
   "devDependencies": {
-    "@types/lodash": "^4.14.191"
+    "@types/lodash": "^4.14.191",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.9"
   }
 }
