+++ frontend/packages/aquascope-editor/src/styles.scss
@@ -91,6 +91,10 @@
     border-color: gainsboro;
     border-radius: 4px;
     color: black;
+
+    margin-bottom: 5px;
+    position: relative;
+    bottom: 3px;
   }
 
   table.perm-step-table {
