+++ frontend/packages/aquascope-embed/package.json
@@ -5,6 +5,8 @@
     "@codemirror/state": "^6.1.2",
     "@codemirror/view": "^6.4.0",
     "@lezer/highlight": "^1.1.3",
-    "aquascope-editor": "workspace:^"
+    "aquascope-editor": "workspace:^",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   }
 }
