+++ frontend/packages/aquascope-standalone/index.html
@@ -18,7 +18,7 @@
     />
     <link
       rel="stylesheet"
-      href="https://use.fontawesome.com/releases/v4.7.0/css/all.css"
+      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.css"
       type="text/css"
     />
     <link rel="stylesheet" type="text/css" href="./styles.css" />
