+++ frontend/packages/aquascope-standalone/package.json
@@ -1,6 +1,5 @@
-{  
+{
   "devDependencies": {
-    "aquascope-editor": "workspace:^",
     "@codemirror/autocomplete": "^6.3.0",
     "@codemirror/commands": "^6.1.2",
     "@codemirror/language": "^6.3.0",
@@ -8,6 +7,9 @@
     "@codemirror/search": "^6.2.2",
     "@codemirror/state": "^6.1.2",
     "@codemirror/view": "^6.4.0",
-    "@replit/codemirror-vim": "^6.0.3"
+    "@replit/codemirror-vim": "^6.0.3",
+    "aquascope-editor": "workspace:^",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   }
 }
